Separate getConfig and getSchema functions into own service
