+++ server/services/index.js
@@ -1,9 +1,11 @@
 'use strict';
 
+const plugin = require( './plugin' );
 const menu = require( './menu' );
 const menuItem = require( './menu-item' );
 
 module.exports = {
   menu,
   'menu-item': menuItem,
+  plugin,
 };
